+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add timeline types to `@scanner` decorator overloads.
+
 ## 0.4.22 (20 March 2026)
 
 - Transcripts: Use `events_data` to reduce memory and storage requirements of `events`.